requirements.txt
PyQt5~=5.15.11
PyPDF2
Pillow
yt-dlp
cryptography
speedtest-cli
requests
beautifulsoup4